requirements.txt
openai==1.28.0
PyPDF2==3.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
rich==13.7.0
